+++ TvSeriesTracker.Web/ClientApp/src/App.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import AppRoutes from './AppRoutes';
-import { Layout } from './components/Layout';
-import './custom.css';
+import { Layout } from './_Layout';
 
 export default class App extends Component {
   static displayName = App.name;
