+++ TvSeriesTracker.Web/ClientApp/src/AppRoutes.js
@@ -1,6 +1,6 @@
-import { Counter } from "./components/Counter";
-import { FetchData } from "./components/FetchData";
-import { Home } from "./components/Home";
+import { Counter } from "./pages/Counter";
+import { FetchData } from "./pages/FetchData";
+import { Home } from "./pages/Home";
 
 const AppRoutes = [
   {
