+++ TvSeriesTracker.Web/ClientApp/src/_Layout.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Container } from 'reactstrap';
-import { NavMenu } from './NavMenu';
+import { NavMenu } from './components/NavMenu';
+import './_Layout.css';
 
 export class Layout extends Component {
   static displayName = Layout.name;
