+++ TvSeriesTracker.Web/ClientApp/src/pages/Counter.js
