+++ TvSeriesTracker.Web/ClientApp/src/pages/FetchData.js
@@ -52,7 +52,7 @@ export class FetchData extends Component {
   }
 
   async populateWeatherData() {
-    const response = await fetch('weatherforecast');
+    const response = await fetch('/api/weatherforecast');
     const data = await response.json();
     this.setState({ forecasts: data, loading: false });
   }
