+++ TvSeriesTracker.Web/ClientApp/src/shared/service-worker.js
