+++ TvSeriesTracker.Web/ClientApp/src/shared/serviceWorkerRegistration.js
