+++ TvSeriesTracker.Web/Controllers/WeatherForecastController.cs
@@ -4,7 +4,7 @@ using TvSeriesTracker.Web.Models;
 namespace TvSeriesTracker.Web.Controllers;
 
 [ApiController]
-[Route("[controller]")]
+[Route("api/[controller]")]
 public class WeatherForecastController : ControllerBase
 {
     private static readonly string[] Summaries = new[]
